% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arcgischeck}
\alias{arcgischeck}
\title{Helper function for the attached shiny function. ArcGIS exports rastered data inverted, this function
aligns the data to the usual grid.}
\usage{
arcgischeck(data, isarcgis)
}
\arguments{
\item{data}{2D data as used in the PALM classes}

\item{isarcgis}{Logical, if data was rastered with ArcGIS set to TRUE}
}
\value{
Returns 2D data, if isarcgis == TRUE, y-axis will be reversed
}
\description{
Helper function for the attached shiny function. ArcGIS exports rastered data inverted, this function
aligns the data to the usual grid.
}
\examples{
a <- matrix(c(1:4),2,2)
b <- arcgischeck(a,TRUE)
all(a!=b)
}
