% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_calc_single_tree.R
\name{f.calc_single_tree}
\alias{f.calc_single_tree}
\title{Calculates the LAD and BAD of a single tree!}
\usage{
f.calc_single_tree(tree_type_b = NULL, tree_shape = NULL,
  crown_ratio = NULL, crown_diameter = NULL, tree_height = NULL,
  lai = NULL, dbh = NULL, extinktion_sphere = 0.6,
  extinktion_cone = 0.2, fillvalue = -9999.9, dx)
}
\arguments{
\item{tree_type_b}{Integer. TreeType according to a "database" so far, only fillvalues!}

\item{tree_shape}{Integer. TreeShape, can be set (1-6), or database, then according to TreeType}

\item{crown_ratio}{Numeric. Ratio of the Crown?}

\item{crown_diameter}{Numeric. Diameter of the Crown, MUST BE SET as database has 0!}

\item{tree_height}{Numeric. Database entry is 12}

\item{lai}{Numeric. Database entry is 2}

\item{dbh}{Numeric. Database entry is 0.35}

\item{extinktion_sphere}{Numeric. ?}

\item{extinktion_cone}{Numeric. =}

\item{fillvalue}{Numeric. -9999.9}
}
\value{
A list with a 3D LAD array, a 3D BAD array and the species of the tree
}
\description{
Calculates the LAD and BAD of a single tree!
}
\examples{
f.calc_single_tree(tree_type_b = 1, crown_diameter = 10, dx = 2)
}
