% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_calc_patch_tree.R
\name{f.calc_tree_patch}
\alias{f.calc_tree_patch}
\title{Calculate Patch LAD fpr PALM-4U}
\usage{
f.calc_tree_patch(lai = array(), canopy_height = array(), dz,
  alpha = 5, beta = 3, fillvalue = -9999.9)
}
\arguments{
\item{lai}{Array. Containts Leaf Area Index for respective patches.}

\item{dz}{Integer. Grid dimension in y direction}

\item{alpha}{Numeric. A Parameter for LAD calculation. See source.}

\item{beta}{Numeric. A Parameter for LAD calculation. See Source.}

\item{fillvalue}{Numeric. Fillvalue for the array}

\item{canopy_heigh}{Array. Contains tree height for respective patches.}
}
\value{
Array. Filled with LAD for each tree patch.
}
\description{
Calculate Patch LAD fpr PALM-4U
}
\examples{
lai <- array(2, c(5,5))
canopy <- array(10, c(5,5))
dz <- 2
lad <- f.calc_tree_patch(lai, canopy, dz)
}
