% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{get_variable_from_wtk}
\alias{get_variable_from_wtk}
\title{Title}
\usage{
get_variable_from_wtk(
  search_string,
  wkt_text,
  numeric = TRUE,
  trailing_comma = TRUE
)
}
\arguments{
\item{search_string}{Character. Which parameter to search in wkt string}

\item{wkt_text}{Character. Wkt string}

\item{numeric}{Boolean. Whether value is numeric or not}
}
\value{

}
\description{
Title
}
