% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_dynamic.R
\name{palm_dynamic}
\alias{palm_dynamic}
\title{Title}
\description{
Title

Title
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{palm_dynamic$new()}}
\item \href{#method-set_initial_profile_lod1}{\code{palm_dynamic$set_initial_profile_lod1()}}
\item \href{#method-set_time}{\code{palm_dynamic$set_time()}}
\item \href{#method-set_forcing}{\code{palm_dynamic$set_forcing()}}
\item \href{#method-export_dynamic_driver}{\code{palm_dynamic$export_dynamic_driver()}}
\item \href{#method-create_soil}{\code{palm_dynamic$create_soil()}}
\item \href{#method-create_dummy_vars}{\code{palm_dynamic$create_dummy_vars()}}
\item \href{#method-prepopoulate_boundaries}{\code{palm_dynamic$prepopoulate_boundaries()}}
\item \href{#method-set_vertical_profile}{\code{palm_dynamic$set_vertical_profile()}}
\item \href{#method-shift_wind_by_topography}{\code{palm_dynamic$shift_wind_by_topography()}}
\item \href{#method-clone}{\code{palm_dynamic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_dynamic$new(dx, nx, ny, nz)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nz}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_initial_profile_lod1"></a>}}
\if{latex}{\out{\hypertarget{method-set_initial_profile_lod1}{}}}
\subsection{Method \code{set_initial_profile_lod1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_dynamic$set_initial_profile_lod1(variable, data.vector)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_time"></a>}}
\if{latex}{\out{\hypertarget{method-set_time}{}}}
\subsection{Method \code{set_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_dynamic$set_time(duration, steps)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_forcing"></a>}}
\if{latex}{\out{\hypertarget{method-set_forcing}{}}}
\subsection{Method \code{set_forcing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_dynamic$set_forcing(variable, data, boundary_side, dim_of_data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_dynamic_driver"></a>}}
\if{latex}{\out{\hypertarget{method-export_dynamic_driver}{}}}
\subsection{Method \code{export_dynamic_driver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_dynamic$export_dynamic_driver(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_soil"></a>}}
\if{latex}{\out{\hypertarget{method-create_soil}{}}}
\subsection{Method \code{create_soil()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_dynamic$create_soil(
  data.vector,
  variable,
  depths = c(0.005, 0.02, 0.06, 0.18, 0.54, 1.62, 4.86, 14.58)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_dummy_vars"></a>}}
\if{latex}{\out{\hypertarget{method-create_dummy_vars}{}}}
\subsection{Method \code{create_dummy_vars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_dynamic$create_dummy_vars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepopoulate_boundaries"></a>}}
\if{latex}{\out{\hypertarget{method-prepopoulate_boundaries}{}}}
\subsection{Method \code{prepopoulate_boundaries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_dynamic$prepopoulate_boundaries(pt = 300, qv = 0.001, u = 0, v = 0, w = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_vertical_profile"></a>}}
\if{latex}{\out{\hypertarget{method-set_vertical_profile}{}}}
\subsection{Method \code{set_vertical_profile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_dynamic$set_vertical_profile(var, data, dim_of_data = "z")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shift_wind_by_topography"></a>}}
\if{latex}{\out{\hypertarget{method-shift_wind_by_topography}{}}}
\subsection{Method \code{shift_wind_by_topography()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_dynamic$shift_wind_by_topography(zt_array = NULL, static_path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_dynamic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
