% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_global.R
\name{palm_global}
\alias{palm_global}
\title{Create global parameters for static driver}
\description{
Create global parameters for static driver

Create global parameters for static driver
}
\examples{

## ------------------------------------------------
## Method `palm_global$new`
## ------------------------------------------------

palm_global$new('Testglobal', 'Max Mustermann', 'Some Institute',
'Somewehere', 1, 1, 0, 0, 33,33, 10)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{palm_global$new()}}
\item \href{#method-changeVar}{\code{palm_global$changeVar()}}
\item \href{#method-changeTimeVar}{\code{palm_global$changeTimeVar()}}
\item \href{#method-clone}{\code{palm_global$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an R6 class that contains all global attributes needed
in the static driver.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_global$new(
  title,
  author,
  institute,
  location,
  x0,
  y0,
  z0,
  t0,
  lat,
  lon,
  dx = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{Title of the file}

\item{\code{author}}{Author, i.e. you}

\item{\code{institute}}{Your institute}

\item{\code{location}}{Site of the static driver}

\item{\code{x0}}{x-coordinate of the lower left point in your coordinate system}

\item{\code{y0}}{y-coordinate of the lower left point in your coordinate system}

\item{\code{z0}}{lowest z-coordinate in your domain}

\item{\code{t0}}{good question}

\item{\code{lat}}{Latitude of the lower left point of your domain}

\item{\code{lon}}{Longitude of the lower left point of your domain}

\item{\code{dx}}{Grid spacing}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Creates the PALM-Header Class
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{palm_global$new('Testglobal', 'Max Mustermann', 'Some Institute',
'Somewehere', 1, 1, 0, 0, 33,33, 10)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-changeVar"></a>}}
\if{latex}{\out{\hypertarget{method-changeVar}{}}}
\subsection{Method \code{changeVar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_global$changeVar(variable, input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-changeTimeVar"></a>}}
\if{latex}{\out{\hypertarget{method-changeTimeVar}{}}}
\subsection{Method \code{changeTimeVar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_global$changeTimeVar(variable, input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_global$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
