% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_berlin_old.R
\name{palm_ncdf_berlin_old}
\alias{palm_ncdf_berlin_old}
\title{Funktion die mit $new aufgerufen wird!
Title}
\description{
Funktion die mit $new aufgerufen wird!
Title

Funktion die mit $new aufgerufen wird!
Title
}
\examples{

## ------------------------------------------------
## Method `palm_ncdf_berlin_old$new`
## ------------------------------------------------

#berlin_headm <- palm_global$new(title = "GIT Example",
#author = "sest",
#institute = "IBP",
#location = "Hoki",
#x0 = 0,   # only important for visualization on a map later on
#y0 = 0,   # only important for visualization on a map later on
#z0 = 0,
#t0 = "2018-06-21 12:00:00 +00", # Character with Date in this format,
# might be important to have correct in later releases of PALM!
#lat = 52.502302,   # important for solar radiation
#lon = 13.364862,   # important for solar radiation
#dx = 5)

berlin_example <- palm_ncdf_berlin$new(filename = "berlin_static.nc",
                   headclass = "berlin_head",
                   pathtofiles = "Path/to/Berlin/data",
                   oldversion = FALSE)
Title


## ------------------------------------------------
## Method `palm_ncdf_berlin_old$importfiles`
## ------------------------------------------------

berlin_example$importfiles(lengthx = 100,
                           lengthy = 100,
                           dx = 10)
Title

Exports a static driver with the data present in the class


## ------------------------------------------------
## Method `palm_ncdf_berlin_old$exportncdf`
## ------------------------------------------------

berlin_example$exportncdf(Path = "Path/to/be/saved"
                          EPSGCode = "EPSG:25833")
Calls ggplot to plot available data in a 2D plot. Contains simple color
coding for pavement, water, buildings and a distinction for small vegetation
and trees. Can display the whole domain or small subareas.
Missing Data is displayed as white and will be marked by the legend.


## ------------------------------------------------
## Method `palm_ncdf_berlin_old$plot_area`
## ------------------------------------------------

berlin_example$plot_area(1,1,200,200)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{palm_ncdf_berlin_old$new()}}
\item \href{#method-importfiles}{\code{palm_ncdf_berlin_old$importfiles()}}
\item \href{#method-exportncdf}{\code{palm_ncdf_berlin_old$exportncdf()}}
\item \href{#method-plot_area}{\code{palm_ncdf_berlin_old$plot_area()}}
\item \href{#method-createofficebuilding}{\code{palm_ncdf_berlin_old$createofficebuilding()}}
\item \href{#method-createresidentialbuilding}{\code{palm_ncdf_berlin_old$createresidentialbuilding()}}
\item \href{#method-createlod1area}{\code{palm_ncdf_berlin_old$createlod1area()}}
\item \href{#method-createbuilding3D}{\code{palm_ncdf_berlin_old$createbuilding3D()}}
\item \href{#method-quickplot}{\code{palm_ncdf_berlin_old$quickplot()}}
\item \href{#method-correct_surface_fraction}{\code{palm_ncdf_berlin_old$correct_surface_fraction()}}
\item \href{#method-fill_areal}{\code{palm_ncdf_berlin_old$fill_areal()}}
\item \href{#method-generate_lai_array}{\code{palm_ncdf_berlin_old$generate_lai_array()}}
\item \href{#method-import_data}{\code{palm_ncdf_berlin_old$import_data()}}
\item \href{#method-SortOverlayingdata}{\code{palm_ncdf_berlin_old$SortOverlayingdata()}}
\item \href{#method-countemptyfields}{\code{palm_ncdf_berlin_old$countemptyfields()}}
\item \href{#method-addsoilandsurfacefraction}{\code{palm_ncdf_berlin_old$addsoilandsurfacefraction()}}
\item \href{#method-add_lod2_variable}{\code{palm_ncdf_berlin_old$add_lod2_variable()}}
\item \href{#method-print}{\code{palm_ncdf_berlin_old$print()}}
\item \href{#method-downscale_resolution}{\code{palm_ncdf_berlin_old$downscale_resolution()}}
\item \href{#method-cutout_static}{\code{palm_ncdf_berlin_old$cutout_static()}}
\item \href{#method-add_any2D_variable}{\code{palm_ncdf_berlin_old$add_any2D_variable()}}
\item \href{#method-clone}{\code{palm_ncdf_berlin_old$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$new(filename, headclass, pathtofiles, oldversion = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Filename}

\item{\code{headclass}}{Variable of the global Class}

\item{\code{pathtofiles}}{Path to the Berlin data provided by the PALM4U Project}

\item{\code{oldversion}}{Leave at FALSE, only set to TRUE if you want to use a PALM version ~ < r2800}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Creates an R6 Class
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#berlin_headm <- palm_global$new(title = "GIT Example",
#author = "sest",
#institute = "IBP",
#location = "Hoki",
#x0 = 0,   # only important for visualization on a map later on
#y0 = 0,   # only important for visualization on a map later on
#z0 = 0,
#t0 = "2018-06-21 12:00:00 +00", # Character with Date in this format,
# might be important to have correct in later releases of PALM!
#lat = 52.502302,   # important for solar radiation
#lon = 13.364862,   # important for solar radiation
#dx = 5)

berlin_example <- palm_ncdf_berlin$new(filename = "berlin_static.nc",
                   headclass = "berlin_head",
                   pathtofiles = "Path/to/Berlin/data",
                   oldversion = FALSE)
Title

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importfiles"></a>}}
\if{latex}{\out{\hypertarget{method-importfiles}{}}}
\subsection{Method \code{importfiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$importfiles(lengthx, lengthy, dx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lengthx}}{Gridpoints in x direction}

\item{\code{lengthy}}{Gridpoints in y direction}

\item{\code{dx}}{Grid spacing}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Imports data into PALM class for relevant PALM-4U data.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{berlin_example$importfiles(lengthx = 100,
                           lengthy = 100,
                           dx = 10)
Title

Exports a static driver with the data present in the class

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exportncdf"></a>}}
\if{latex}{\out{\hypertarget{method-exportncdf}{}}}
\subsection{Method \code{exportncdf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$exportncdf(Path = self$path, EPSGCode = "EPSG:25833")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Path}}{Path where the driver is to be saved}

\item{\code{EPSGCode}}{Necessary for correct display on maps. Currently only
usable for few coordinate systems}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Exports a static driver
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{berlin_example$exportncdf(Path = "Path/to/be/saved"
                          EPSGCode = "EPSG:25833")
Calls ggplot to plot available data in a 2D plot. Contains simple color
coding for pavement, water, buildings and a distinction for small vegetation
and trees. Can display the whole domain or small subareas.
Missing Data is displayed as white and will be marked by the legend.

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_area"></a>}}
\if{latex}{\out{\hypertarget{method-plot_area}{}}}
\subsection{Method \code{plot_area()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$plot_area(xleft, yleft, xl = 40, yl = 40)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xleft}}{Distance to lower left corner in x direction. 1 is coordinate origin}

\item{\code{yleft}}{Distance to lower left corner in y direction. 1 is coordinate origin}

\item{\code{xl}}{Length of plot in x direction. Cannot exceed overall dimensions of domain}

\item{\code{yl}}{Length of plot in y direction. Cannot exceed overall dimensions of domain}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Simple visualization of the simulation domain
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{berlin_example$plot_area(1,1,200,200)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createofficebuilding"></a>}}
\if{latex}{\out{\hypertarget{method-createofficebuilding}{}}}
\subsection{Method \code{createofficebuilding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$createofficebuilding(
  start,
  size,
  height,
  ignorestreets = TRUE,
  notnew = c()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createresidentialbuilding"></a>}}
\if{latex}{\out{\hypertarget{method-createresidentialbuilding}{}}}
\subsection{Method \code{createresidentialbuilding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$createresidentialbuilding(
  start,
  size,
  height,
  ignorestreets = TRUE,
  notnew = c()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createlod1area"></a>}}
\if{latex}{\out{\hypertarget{method-createlod1area}{}}}
\subsection{Method \code{createlod1area()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$createlod1area(start, array2d)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createbuilding3D"></a>}}
\if{latex}{\out{\hypertarget{method-createbuilding3D}{}}}
\subsection{Method \code{createbuilding3D()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$createbuilding3D(force = FALSE, orogrphy3d = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-quickplot"></a>}}
\if{latex}{\out{\hypertarget{method-quickplot}{}}}
\subsection{Method \code{quickplot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$quickplot(variable)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-correct_surface_fraction"></a>}}
\if{latex}{\out{\hypertarget{method-correct_surface_fraction}{}}}
\subsection{Method \code{correct_surface_fraction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$correct_surface_fraction()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_areal"></a>}}
\if{latex}{\out{\hypertarget{method-fill_areal}{}}}
\subsection{Method \code{fill_areal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$fill_areal(
  variabletofill,
  startx,
  starty,
  fillvalue,
  valuetobefilled = NULL,
  boundaryarea = NULL,
  overwrite = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_lai_array"></a>}}
\if{latex}{\out{\hypertarget{method-generate_lai_array}{}}}
\subsection{Method \code{generate_lai_array()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$generate_lai_array(
  dz,
  fixed_tree_height = NULL,
  alpha = 5,
  beta = 3,
  startx = NULL,
  starty = NULL,
  lengthx = NULL,
  lengthy = NULL,
  additional_array = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import_data"></a>}}
\if{latex}{\out{\hypertarget{method-import_data}{}}}
\subsection{Method \code{import_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$import_data(v.file, palmtype, typeid, street = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SortOverlayingdata"></a>}}
\if{latex}{\out{\hypertarget{method-SortOverlayingdata}{}}}
\subsection{Method \code{SortOverlayingdata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$SortOverlayingdata(inorderof = "BPWV")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-countemptyfields"></a>}}
\if{latex}{\out{\hypertarget{method-countemptyfields}{}}}
\subsection{Method \code{countemptyfields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$countemptyfields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addsoilandsurfacefraction"></a>}}
\if{latex}{\out{\hypertarget{method-addsoilandsurfacefraction}{}}}
\subsection{Method \code{addsoilandsurfacefraction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$addsoilandsurfacefraction(type_soil = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_lod2_variable"></a>}}
\if{latex}{\out{\hypertarget{method-add_lod2_variable}{}}}
\subsection{Method \code{add_lod2_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$add_lod2_variable(name, v.data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downscale_resolution"></a>}}
\if{latex}{\out{\hypertarget{method-downscale_resolution}{}}}
\subsection{Method \code{downscale_resolution()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$downscale_resolution(factor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cutout_static"></a>}}
\if{latex}{\out{\hypertarget{method-cutout_static}{}}}
\subsection{Method \code{cutout_static()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$cutout_static(startp, endp, sure = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_any2D_variable"></a>}}
\if{latex}{\out{\hypertarget{method-add_any2D_variable}{}}}
\subsection{Method \code{add_any2D_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$add_any2D_variable(
  variablename,
  variable_list,
  print_template = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_berlin_old$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
