% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_ncdf_data_template.R
\name{palm_ncdf_data_template}
\alias{palm_ncdf_data_template}
\title{Title}
\description{
Title

Title
}
\examples{

## ------------------------------------------------
## Method `palm_ncdf_data_template$new`
## ------------------------------------------------

x <- 1
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{palm_ncdf_data_template$new()}}
\item \href{#method-add_another_data}{\code{palm_ncdf_data_template$add_another_data()}}
\item \href{#method-print}{\code{palm_ncdf_data_template$print()}}
\item \href{#method-clone}{\code{palm_ncdf_data_template$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_data_template$new(
  dat_name,
  FillValue,
  d_units,
  longname,
  d_source,
  lod,
  data,
  d_type,
  dimensions
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat_name}}{Variable Name for StaticDriver}

\item{\code{FillValue}}{Fill_Value of Data}

\item{\code{d_units}}{Units of Data}

\item{\code{longname}}{LongName}

\item{\code{d_source}}{Source of Data}

\item{\code{lod}}{Level of Detail}

\item{\code{data}}{Data!}

\item{\code{d_type}}{DataType}

\item{\code{dimensions}}{Which Dimensions correspond to data?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A R6 Class with the needed Setup for the palm_ncdf_berlin clones.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- 1
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_another_data"></a>}}
\if{latex}{\out{\hypertarget{method-add_another_data}{}}}
\subsection{Method \code{add_another_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_data_template$add_another_data(
  dat_name,
  FillValue,
  d_units,
  longname,
  d_source,
  lod,
  data,
  d_type,
  dimensions
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_data_template$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_data_template$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
