% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_ncdf_dimension_template.R
\name{palm_ncdf_dimension_template}
\alias{palm_ncdf_dimension_template}
\title{Title}
\description{
Title

Title
}
\examples{

## ------------------------------------------------
## Method `palm_ncdf_dimension_template$new`
## ------------------------------------------------

x <- 1
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{palm_ncdf_dimension_template$new()}}
\item \href{#method-add_another_dim}{\code{palm_ncdf_dimension_template$add_another_dim()}}
\item \href{#method-print}{\code{palm_ncdf_dimension_template$print()}}
\item \href{#method-clone}{\code{palm_ncdf_dimension_template$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_dimension_template$new(longname, standardname, units, vals)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{longname}}{Longname of the Dimension}

\item{\code{standardname}}{Standardname of the Dimension, also used as VarName in ncdf!}

\item{\code{units}}{Units of the Dimensions}

\item{\code{vals}}{Values of the Dimensions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A R6 Class, that resembles a Dimension in ncdf!
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- 1
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_another_dim"></a>}}
\if{latex}{\out{\hypertarget{method-add_another_dim}{}}}
\subsection{Method \code{add_another_dim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_dimension_template$add_another_dim(
  longname,
  standardname,
  units,
  vals
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_dimension_template$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{palm_ncdf_dimension_template$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
