% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_import_static.R
\docType{data}
\name{palm_ncdf_import}
\alias{palm_ncdf_import}
\title{R6-Class to import data from an already existing static driver}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
palm_ncdf_import
}
\arguments{
\item{newfilename}{Character. Filename for the future static driver}

\item{pathtofiles}{Character. Path where the old static driver lies}

\item{oldfilename}{Character. Name of the old static driver}

\item{oldversion}{Logical. Only use TRUE for PALM versions <2900}
}
\value{
The initialize/new function creates a barebones class with references to the static driver.
To import the data of the static file, read_static() has to be called afterwards
}
\description{
R6-Class to import data from an already existing static driver
}
\examples{
import_static <- palm_ncdf_import$new("new_static_driver.nc",
                                      "Path/to/files", "old_static_driver.nc", F)
import_static$read_static()
}
\keyword{datasets}
