% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_manual_grid.R
\docType{data}
\name{palm_ncdf_manual}
\alias{palm_ncdf_manual}
\title{R6-Class to create a new static driver from scratch. Creates the most important data arrays depending on the
grid definitions, given by the initialize/new function}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
palm_ncdf_manual
}
\arguments{
\item{nx}{Numerical. Grid points in x direction}

\item{ny}{Numerical. Grid points in y direction}

\item{dx}{Numerical. Grid spacing}

\item{headclass}{R6-Class.}

\item{oldversion}{Logical. Only use TRUE for PALM versions <2900}

\item{gui.arcgis}{Logical. Set to TRUE, when you want to import rastered files created by ARCGIS.}
}
\value{
The initialize/new functions creates a R6 class, where most important arrays (vegetation_type, buildings_2d, etc)
are initialized with fill values.
}
\description{
R6-Class to create a new static driver from scratch. Creates the most important data arrays depending on the
grid definitions, given by the initialize/new function
}
\examples{
manual_static <- palm_ncdf_manual$new(200,200,10, palm_headclass, FALSE, FALSE)
}
\keyword{datasets}
