% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_shiny.R
\docType{data}
\name{palm_ncdf_shiny}
\alias{palm_ncdf_shiny}
\title{R6 Class to create a static driver for the shiny app. The domain is defined via the topography file, other
data arrays have to be defined later.}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
palm_ncdf_shiny
}
\arguments{
\item{topofile}{Link to a rastered ncfile with topography. May not conatin fill values!}

\item{headclass}{R6-Class.}

\item{oldversion}{Logical. Only use TRUE for PALM versions <2900}

\item{gui.arcgis}{Logical. Set to TRUE, when you want to import rastered files created by ARCGIS.}
}
\value{
The initialize/new functions creates a R6 class that only contains the array "zt". Other information has to be
loaded via import_files()
}
\description{
R6 Class to create a static driver for the shiny app. The domain is defined via the topography file, other
data arrays have to be defined later.
}
\examples{
shiny_static <- palm_ncdf_shiny$new("Path/to/topfile.nc", palm_headclass, FALSE, FALSE)
}
\keyword{datasets}
