% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palm_ncdf_berlin.R
\name{plot.palm_ncdf_berlin}
\alias{plot.palm_ncdf_berlin}
\title{Title}
\usage{
\method{plot}{palm_ncdf_berlin}(obj)
}
\arguments{
\item{obj}{A R6 plalm_ncdf Class}
}
\value{
A ggplot
}
\description{
Title
}
\examples{
a <- 1
}
