% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{plot_2d_data}
\alias{plot_2d_data}
\title{Title}
\usage{
plot_2d_data(matrix)
}
\arguments{
\item{matrix}{A 2D Matrix}
}
\value{
Returns a ggplot
}
\description{
Title
}
\examples{
x <- matrix(runif(100), ncol = 10, nrow =10 )
plot_2d_data(x)
}
