% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pint_palm_ncdf_berlin.R
\name{print.palm_ncdf_berlin}
\alias{print.palm_ncdf_berlin}
\title{Title}
\usage{
\method{print}{palm_ncdf_berlin}(obj)
}
\arguments{
\item{obj}{Call print function of R6 palm_ncdf class}
}
\value{
Some Strings
}
\description{
Title
}
