% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R, R/utils.R
\name{specify_decimal}
\alias{specify_decimal}
\title{Helper function that specifies trailing positions of decimals}
\usage{
specify_decimal(x, k)

specify_decimal(x, k)
}
\arguments{
\item{x}{Any numeric}

\item{k}{Trailing positions of numeric}

\item{x}{Any numeric}

\item{k}{Trailing positions of numeric}
}
\value{
A numeric with the chosen trailing positions

A numeric with the chosen trailing positions
}
\description{
Helper function that specifies trailing positions of decimals
}
\examples{
specify_decimal( 3.14159265359, 3)
specify_decimal( 3.14159265359, 3)
}
