% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{substrRight}
\alias{substrRight}
\title{Generic function to cut off n characters from a string}
\usage{
substrRight(x, n)

substrRight(x, n)
}
\arguments{
\item{x}{A string of characters}

\item{n}{Amount of characters to be returned}

\item{x}{A string of characters}

\item{n}{Amount of characters to be returned}
}
\value{
The cut off string.

The cut off string.
}
\description{
Generic function to cut off n characters from a string

Generic function to cut off n characters from a string
}
\examples{
a <- substrRight("R_is_awesome", 7)
a
a <- substrRight("Awesome_is_R", 7)
a
}
