% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_berlin_old.R
\docType{data}
\name{palm_ncdf_berlin_old}
\alias{palm_ncdf_berlin_old}
\title{Funktion die mit $new aufgerufen wird!
Title}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
palm_ncdf_berlin_old
}
\arguments{
\item{filename}{Filename}

\item{headclass}{Variable of the global Class}

\item{pathtofiles}{Path to the Berlin data provided by the PALM4U Project}

\item{oldversion}{Leave at FALSE, only set to TRUE if you want to use a PALM version ~ < r2800}

\item{lengthx}{Gridpoints in x direction}

\item{lengthy}{Gridpoints in y direction}

\item{dx}{Grid spacing}

\item{Path}{Path where the driver is to be saved}

\item{EPSGCode}{Necessary for correct display on maps. Currently only
usable for few coordinate systems}

\item{xleft}{Distance to lower left corner in x direction. 1 is coordinate origin}

\item{yleft}{Distance to lower left corner in y direction. 1 is coordinate origin}

\item{xl}{Length of plot in x direction. Cannot exceed overall dimensions of domain}

\item{yl}{Length of plot in y direction. Cannot exceed overall dimensions of domain}
}
\value{
Creates an R6 Class

Imports data into PALM class for relevant PALM-4U data.

Exports a static driver

Simple visualization of the simulation domain
}
\description{
Funktion die mit $new aufgerufen wird!
Title
}
\examples{
#berlin_headm <- palm_global$new(title = "GIT Example",
#author = "sest",
#institute = "IBP",
#location = "Hoki",
#x0 = 0,   # only important for visualization on a map later on
#y0 = 0,   # only important for visualization on a map later on
#z0 = 0,
#t0 = "2018-06-21 12:00:00 +00", # Character with Date in this format,
# might be important to have correct in later releases of PALM!
#lat = 52.502302,   # important for solar radiation
#lon = 13.364862,   # important for solar radiation
#dx = 5)

berlin_example <- palm_ncdf_berlin$new(filename = "berlin_static.nc",
                   headclass = "berlin_head",
                   pathtofiles = "Path/to/Berlin/data",
                   oldversion = FALSE)
Title

berlin_example$importfiles(lengthx = 100,
                           lengthy = 100,
                           dx = 10)
Title

Exports a static driver with the data present in the class

berlin_example$exportncdf(Path = "Path/to/be/saved"
                          EPSGCode = "EPSG:25833")
Calls ggplot to plot available data in a 2D plot. Contains simple color
coding for pavement, water, buildings and a distinction for small vegetation
and trees. Can display the whole domain or small subareas.
Missing Data is displayed as white and will be marked by the legend.

berlin_example$plot_area(1,1,200,200)
}
\keyword{datasets}
