% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_ncdf_data_template.R
\docType{data}
\name{palm_ncdf_data_template}
\alias{palm_ncdf_data_template}
\title{Title}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
palm_ncdf_data_template
}
\arguments{
\item{dat_name}{Variable Name for StaticDriver}

\item{FillValue}{Fill_Value of Data}

\item{d_units}{Units of Data}

\item{longname}{LongName}

\item{d_source}{Source of Data}

\item{lod}{Level of Detail}

\item{data}{Data!}

\item{d_type}{DataType}

\item{dimensions}{Which Dimensions correspond to data?}
}
\value{
A R6 Class with the needed Setup for the palm_ncdf_berlin clones.
}
\description{
Title
}
\examples{
x <- 1
}
\keyword{datasets}
