% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_ncdf_dimension_template.R
\docType{data}
\name{palm_ncdf_dimension_template}
\alias{palm_ncdf_dimension_template}
\title{Title}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
palm_ncdf_dimension_template
}
\arguments{
\item{longname}{Longname of the Dimension}

\item{standardname}{Standardname of the Dimension, also used as VarName in ncdf!}

\item{units}{Units of the Dimensions}

\item{vals}{Values of the Dimensions}
}
\value{
A R6 Class, that resembles a Dimension in ncdf!
}
\description{
Title
}
\examples{
x <- 1
}
\keyword{datasets}
